CRUD

+++ src/main/java/com/example/springbootapelsin/service/CustomerService.java
@@ -5,6 +5,7 @@ import com.example.springbootapelsin.entity.Customer;
 import com.example.springbootapelsin.repository.CategoryRepository;
 import com.example.springbootapelsin.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 
@@ -28,21 +29,23 @@ public class CustomerService {
         model.addAttribute("list",customers);
     }
 
-    public void edit(Integer id, Customer customer, Model model) {
+    public void edit(Integer id, Model model, Customer customer) {
         Optional<Customer> optionalCustomer = customerRepository.findById(id);
         if (optionalCustomer.isPresent()) {
-           Customer editedCustomer = optionalCustomer.get();
+            Customer editedCustomer = optionalCustomer.get();
             editedCustomer.setName(customer.getName());
             editedCustomer.setCountry(customer.getCountry());
             editedCustomer.setAddress(customer.getAddress());
             editedCustomer.setPhone(customer.getPhone());
             customerRepository.save(editedCustomer); // malumot qo'shish
         }
-        model.addAttribute("list", customerRepository.findAll());
+        model.addAttribute("list", customerRepository.findAll(Sort.by("id")));
     }
-
     public void delete(Integer id, Model model) {
-        customerRepository.deleteById(id);
-        model.addAttribute("list",customerRepository.findAll());
+        Optional<Customer> optionalCustomer = customerRepository.findById(id);
+        if (optionalCustomer.isPresent()) {
+            customerRepository.deleteById(id);
+        }
+        model.addAttribute("list", customerRepository.findAll(Sort.by("id")));
     }
 }
